% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.ls_optim}
\alias{coef.ls_optim}
\title{Coefficients for \code{ls_optim}}
\usage{
\method{coef}{ls_optim}(object, ...)
}
\arguments{
\item{object}{An \code{"ls_optim"} fit.}

\item{...}{Unused.}
}
\value{
Named numeric vector of coefficients.
}
\description{
Coefficients for \code{ls_optim}
}
