% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{expit}
\alias{expit}
\title{Expit/sigmoid function}
\usage{
expit(x)
}
\arguments{
\item{x}{is a vector with probability values
expit()}
}
\description{
This function used to compute the logit values for logistic regression function
}
