% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fitted.ls_optim}
\alias{fitted.ls_optim}
\title{Fitted values for \code{ls_optim}}
\usage{
\method{fitted}{ls_optim}(object, ...)
}
\arguments{
\item{object}{An \code{"ls_optim"} fit.}

\item{...}{Unused.}
}
\value{
Numeric vector of fitted values.
}
\description{
Fitted values for \code{ls_optim}
}
