% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{myLasso}
\alias{myLasso}
\title{Lasso solution path Function}
\usage{
myLasso(X, Y, lambda_all)
}
\arguments{
\item{X}{An n x p matrix of explanatory variables}

\item{Y}{An n dimensional response vector}

\item{lambda_all}{A vector of regularization parameters (Make sure to sort lambda_all in decreasing order for efficiency)}
}
\description{
This function allows you to find the lasso solution path for various values of the regularization parameter lambda.
Function should return a matrix containing the lasso solution vector beta for each regularization parameter.
}
