% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{myLM}
\alias{myLM}
\title{Linear Regression Function}
\usage{
myLM(X, Y)
}
\arguments{
\item{X}{An n x p matrix of explanatory variables}

\item{Y}{An n dimensional vector of responses}
}
\description{
This function allows you to do linear regression with intercept by QR decomposition.
Function returns the least squares solution vector.
}
