% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{myLM_0}
\alias{myLM_0}
\title{Linear Regression Intercept Free Function}
\usage{
myLM_0(X, Y)
}
\arguments{
\item{X}{An n x p matrix of explanatory variables}

\item{Y}{An n dimensional vector of responses}
}
\description{
This function allows you to do linear regression without intercept by QR decomposition.
Function returns the least squares solution vector.
}
