% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{myLogistic}
\alias{myLogistic}
\title{Logistic Regression Function}
\usage{
myLogistic(X, Y)
}
\arguments{
\item{X}{An n x p matrix of explanatory variables}

\item{Y}{An n dimensional vector of binary responses}
}
\description{
This function allows you to do Logistic regression with binomial family without intercept by solving iterated reweighed least squares.
Function returns the logistic regression solution vector.
}
