% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{myPCA}
\alias{myPCA}
\title{PCA Function}
\usage{
myPCA(A, numIter = 1000)
}
\arguments{
\item{A}{A square matrix}

\item{numIter}{Number of iterations}
}
\description{
This function allows you to perfrom PCA by using the QR function.
Function should output a list with D and V.
}
