% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{myQR}
\alias{myQR}
\title{QR Function}
\usage{
myQR(A)
}
\arguments{
\item{A}{The given matrix to do QR decomposition with}
}
\description{
This function allows you to do QR decomposition on the given matrix.
The output Q and R should satisfy: A=Q%*%R.
}
