% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{myRidge}
\alias{myRidge}
\title{Ridge Regression Function}
\usage{
myRidge(X, Y, lambda)
}
\arguments{
\item{X}{An n x p matrix of explanatory variables}

\item{Y}{An n vector of dependent variables or a matrix as long as the function works}

\item{lambda}{Regularization parameter (lambda >= 0)}
}
\description{
This function allows you to perfrom ridge regression of Y on X.
Function should return beta, the ridge regression solution.
}
