% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{mySpline}
\alias{mySpline}
\title{Spline Regression Function}
\usage{
mySpline(x, Y, lambda, p = 100)
}
\arguments{
\item{x}{An n x 1 vector or matrix of explanatory variables}

\item{Y}{An n x 1 vector of dependent variables or a matrix as long as the function works}

\item{lambda}{Regularization parameter (lambda >= 0)}

\item{p}{Number of cuts to make to the x-axis (default p = 100)}
}
\description{
This function allows you to perfrom spline regression of Y on X.
Function should return a list containing the spline regression beta vector and the predicted Y values.
}
