% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRegression.R
\name{mySweep}
\alias{mySweep}
\title{Sweep Function}
\usage{
mySweep(A, m)
}
\arguments{
\item{A}{A square matrix}

\item{m}{The pivot element is A[m,m]}
}
\description{
This function allows you to perfrom a Sweep operation on A with pivot element A[m,m].
Function should output a swept matrix.
}
