% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.ls_optim}
\alias{predict.ls_optim}
\title{Predict method for \code{ls_optim}}
\usage{
\method{predict}{ls_optim}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An \code{"ls_optim"} fit.}

\item{newdata}{Optional matrix or data frame of predictors (without intercept).
If omitted, in-sample fitted values are returned.}

\item{...}{Unused.}
}
\value{
Numeric vector of predictions.
}
\description{
Predict method for \code{ls_optim}
}
\examples{
\donttest{
set.seed(1)
X <- matrix(rnorm(50*2), 50, 2); colnames(X) <- c("x1","x2")
y <- as.vector(cbind(1, X) \%*\% c(1, 2, -1) + rnorm(50))
fit <- my_lm(X, y, B = 20)
predict(fit, newdata = X[1:3, ])
}

}
