% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.ls_optim}
\alias{print.ls_optim}
\title{Print method for \code{ls_optim}}
\usage{
\method{print}{ls_optim}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ls_optim"}.}

\item{...}{Unused; included for S3 consistency.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for \code{ls_optim}
}
