% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.summary.ls_optim}
\alias{print.summary.ls_optim}
\title{Print method for \code{summary.ls_optim}}
\usage{
\method{print}{summary.ls_optim}(x, digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{x}{An object of class \code{"summary.ls_optim"}.}

\item{digits}{Integer; number of significant digits to print. Defaults to
two fewer than \code{getOption("digits")} but at least 3.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for \code{summary.ls_optim}
}
