% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{residuals.ls_optim}
\alias{residuals.ls_optim}
\title{Residuals for \code{ls_optim}}
\usage{
\method{residuals}{ls_optim}(object, ...)
}
\arguments{
\item{object}{An \code{"ls_optim"} fit.}

\item{...}{Unused.}
}
\value{
Numeric vector of residuals.
}
\description{
Residuals for \code{ls_optim}
}
