% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vcov.ls_optim}
\alias{vcov.ls_optim}
\title{Variance-covariance matrix for \code{ls_optim}}
\usage{
\method{vcov}{ls_optim}(object, ...)
}
\arguments{
\item{object}{An \code{"ls_optim"} fit.}

\item{...}{Unused.}
}
\value{
Numeric covariance matrix of coefficients (bootstrap estimate).
}
\description{
Variance-covariance matrix for \code{ls_optim}
}
