% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_lm}
\alias{my_lm}
\title{Linear regression via optimization}
\usage{
my_lm(
  X,
  y,
  B = 100,
  add_intercept = TRUE,
  optim_method = "BFGS",
  optim_control = list(maxit = 1000),
  ...
)
}
\arguments{
\item{X}{A \code{matrix} (not a data frame) of dimension nxp representing numerical predictors}

\item{y}{A \code{vector} that represents the numeric response.}

\item{B}{A \code{numeric} (integer) used to denote the number of bootstrap replications.}

\item{seed}{A \code{numeric} used to control the seed of the random number
generator used by this function.}
}
\value{
A \code{list} of class ls_optim:
\describe{
     \item{I}{Estimated value of the integral}
     \item{var}{Estimated variance of the estimator}
}
}
\description{
Uses a least-square loss to estimate coefficients and provide inference tools
}
\examples{
set.seed(1)
n <- 120; p <- 8
X <- matrix(rnorm(n * p), n, p)
colnames(X) <- paste0("x", 1:p)
beta_true <- c(2, -1.5, 0, 0, 1, rep(0, p - 5))
y <- as.vector(X \%*\% beta_true + rnorm(n, sd = 1))
fit <- my_lm(X, y, B = 200, add_intercept = TRUE)
}
\author{
Roberto Molinari
}
